chore: simplify cleanups
