+++ packages/workspace-fusion/src/lib/utils/viewController/index.ts
@@ -1,1 +1,1 @@
-export { addViewController, switchTabOnNavigation } from './addViewController';
+export { addViewController } from './addViewController';
